mtd-interface: add UBIATTACH config option
Attaching random MTD devices as a UBI device is potentially dangerous,
because the kernel will start writing to it if it is found to be
erased. This may e.g. destroy a bootloader on it.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
---
Some parts of the code are only relevant when UBIATTACH is enabled but
are not #ifdef'ed away for simplicity. In particular, all the support
code for blacklisting is not relevant.