core: blacklisting should depend on CONFIG_UBIATTACH
The blacklisting options should depend on CONFIG_UBIATTACH instead of
CONFIG_MTD. Indeed, blacklisting is only relevant for attaching a UBI
device. Thus, convert the CONFIG_MTD ifdefs into CONFIG_UBIATTACH in
the relevant locations.

In addition, there were two places where the ifdef was missing:
- in the parsing of swupdate.cfg;
- in the definition of struct swupdate_global_cfg.
So add the ifdef there.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
---
If patch 1/2 is not applied, this patch is still relevant but then
with UBIVOL instead of UBIATTACH. I'll respin as needed.